read_at_offset takes size_t offset

+++ mlx/c/iotypes.h
@@ -37,7 +37,7 @@ typedef struct mlx_io_vtable_ {
   size_t (*tell)(void*);
   void (*seek)(void*, int64_t off, int whence);
   void (*read)(void*, char* data, size_t n);
-  void (*read_at_offset)(void*, char* data, size_t n, int off);
+  void (*read_at_offset)(void*, char* data, size_t n, size_t off);
   void (*write)(void*, const char* data, size_t n);
   const char* (*label)(void*);
 } mlx_io_vtable;
